requirements.txt
streamlit==1.28.1
plotly==5.17.0
numpy==1.24.3
matplotlib>=3.7.0
pandas>=1.5.0
scikit-learn>=1.3.0
# Kafka dependencies for SDV network
kafka-python==2.0.2
cryptography==41.0.7
# CAN bus simulation (optional)
python-can>=4.0.0
# API support
flask>=2.0.0
flask-cors>=4.0.0
requests>=2.25.0
# FastAPI for vehicle security API
fastapi>=0.68.0
uvicorn>=0.15.0
# Analytics support
plotly>=5.0.0
pandas>=1.3.0
# S3 support (optional)
boto3>=1.26.0
# Additional dependencies
psutil>=5.8.0
pytest>=6.0.0
